# [mypyc] Fast tuple equality checks