--- a/mypyc/irbuild/builder.py
+++ b/mypyc/irbuild/builder.py
@@ -192,6 +192,10 @@ class IRBuilder:
         return self.builder.unary_op(lreg, expr_op, line)
 
     def binary_op(self, lreg: Value, rreg: Value, expr_op: str, line: int) -> Value:
+        # special case tuple comparison here so that nested tuples can be supported
+        if (isinstance(lreg.type, RTuple) and isinstance(rreg.type, RTuple)
+                and expr_op in ('==', '!=')):
+            return self.compare_tuples(lreg, rreg, expr_op, line)
         return self.builder.binary_op(lreg, rreg, expr_op, line)
 
     def coerce(self, src: Value, target_type: RType, line: int, force: bool = False) -> Value:
@@ -253,6 +257,9 @@ class IRBuilder:
 
     def compare_tagged(self, lhs: Value, rhs: Value, op: str, line: int) -> Value:
         return self.builder.compare_tagged(lhs, rhs, op, line)
+
+    def compare_tuples(self, lhs: Value, rhs: Value, op: str, line: int) -> Value:
+        return self.builder.compare_tuples(lhs, rhs, op, line)
 
     def builtin_len(self, val: Value, line: int) -> Value:
         return self.builder.builtin_len(val, line)