closes mypyc/mypyc#728