@TH3CHARLie 2020-08-24
some performance numbers(running on my mac locally, very imprecise)

On master
```
interpreted: 0.180333s (avg of 5 iterations; stdev 0.61%)
compiled:    0.210294s (avg of 5 iterations; stdev 0.9%)

compiled is 0.858x faster
```

with this PR:
```
interpreted: 0.182221s (avg of 6 iterations; stdev 0.51%)
compiled:    0.055916s (avg of 6 iterations; stdev 0.35%)

compiled is 3.259x faster
```