@msullivan 2020-08-25
No need for this to be in this PR, but it would be pretty straightforward to do `<` and friends too, right?