@TH3CHARLie 2020-08-25
> No need for this to be in this PR, but it would be pretty straightforward to do `<` and friends too, right?

Sure, only need to do some minor modifications. But they appear much less than these two ops, so probably it's OK to lower their priority IMO.